requirements.txt
PyYAML==5.4.1
black==20.8b1
pandas==1.1.5
pylint==2.6.0
releng-client==0.0.1
